[project @ 2002-03-15 16:07:58 by rje] Add ability for NoFib to be run with various different modes.
The mode is set by changing a variable called "mode".

This allows tests to be run for different lengths of time, depending on what kin
d of test is wanted.

Current modes are:

<unset> = default mode. Like previous NoFib.
"slow" = tries to run for around 10-20 seconds on my workstation.

A test is given several modes by giving it the following:
        <test>.stdout -> <test>.<mode>stdout
        <test>.stdin -> <test>.<mode>stdin

Args are set in the makefile as

<mode>_OPTS