[project @ 2002-03-15 16:14:04 by rje] Extended the imaginary tests so that they can be run in slow mode by setting mode=slow.
Some of these tests required minor changes in order to be able to run at
more than one speed. In particular, gen_regexps now only outputs the number
of chars in its output, rather than all the regular expressions. This is
to avoid the slow stdout file having to be insanely huge.

I've done a half-hearted attempt at making the speculative tests go slowly
too, but that is too incomplete and too messy to be committed just yet.