[project @ 2002-11-07 13:17:43 by simonmar] Make 'make mode=fast' work, and tweak the settings slightly so that all of the tests run in <1s (except exp3_8, which runs in just over 1s), with most of the tests running in about 0.3s.  This is on a 1Gz PIII.
The normal settings are unchanged.  In fact in some cases, the fast
settings are slower than the normal settings... at some point we'll
need to bump all the normal settings too, but that needs to be done in
one large sweep so we invalidate all the old results in one go.