[project @ 2005-04-28 13:34:25 by simonmar] mode=slow
That completes the imaginary and spectral suites.  Most of the
programs in here will now run in 1-2s with mode=slow (on a 3GHz P4).
I had to modify the programs in a few places, so there will be some
slight discrepancies in the size of Main.o due to these changes.