Add the fibon benchmarks
There are a total of 34 benchmarks divided into four subgroups. The
DPH, Repa, and Shootout benchmarks are available elsewhere, but I
went ahead and added them since they are part of the fibon suite and
it might be nice to have a version of them easily available for use
by GHC developers.  As far as I know, the Hackage benchmarks are not
available (as benchmarks) anywhere else.