real: enable linear
This also tweaks a parameter in the benchmark so that it takes a
reasonable amount of time to run ( ~1.1s on my machine) and adds
the `.stdout` file. The `Makefile` required some tweaking to make sure
that the modules are compiled in the right order.

Signed-off-by: Michal Terepeta <michal.terepeta@gmail.com>

Test Plan: compile & run nofib

Reviewers: bgamari

Reviewed By: bgamari

Differential Revision: https://phabricator.haskell.org/D3157