real: remove `rx`
The benchmark doesn't compile, but even when fixed, it doesn't seem
very useful - it runs in mere ~100ms and there aren't easy knobs to
make it run for longer. Considering that this hasn't been used for
some time, it seems ok to simply remove it.

Also, removing it will make the initial version of Shake-based build
system easier.

Signed-off-by: Michal Terepeta <michal.terepeta@gmail.com>

Test Plan: compile & run nofib

Reviewers: goldfire, bgamari

Reviewed By: bgamari

Differential Revision: https://phabricator.haskell.org/D3159