spectral/mate: remove unnecessary HS_SRCS
I've mistakenly added it as a workaround to build problems, but the
whole problem was only due a missing `.depend` file (which is created
by `make boot` and interestigly is not cleaned by `make clean`). In
any way, the `HS_SRCS` is simply unnecessary and can be removed.

Signed-off-by: Michal Terepeta <michal.terepeta@gmail.com>

Test Plan: build & run

Reviewers: bgamari

Reviewed By: bgamari

Differential Revision: https://phabricator.haskell.org/D3389