Disable timer-based context switches
Summary:
In the past, we repeatedly had problems with non-deterministic allocations due to stack squeezing during context switches (#4450, #8611).
This patch adds `+RTS -V0 -RTS` as extra `RUNTEST_OPTS` for every single-threaded benchmark.

Is this the right place to add the flags? Should we also do this for all the other benchmarks?

Reviewers: simonmar, osa1, nomeata, O26 nofib

GHC Trac Issues: #8611

Differential Revision: https://phabricator.haskell.org/D5470