Make `git clean -nxd` silent after `make distclean`
Also check that invariant in CI: After a `make distclean`
`git clean -nxd` should find no files.

The hope is that this catches cases were a `make clean` would forget to
list a `.gitignore`d file that should either be registered as a
`(DIST_)CLEAN_FILE` or should be unignored and committed into the
repository.